+++ flowr/src/cli/cli_debug_handler.rs
@@ -12,7 +12,7 @@ use flowrlib::run_state::{RunState, State};
 use crate::{BlockBreakpoint, CoordinatorConnection, DataBreakpoint, ExecutionEnded, ExecutionStarted,
             ExitingDebugger, JobCompleted, JobError, Panic, PriorToSendingJob, Resetting,
             WaitingForCommand};
-use crate::cli::client_coordinator::WAIT;
+use crate::cli::connections::WAIT;
 use crate::DebugServerMessage::{BlockState, Error, FlowUnblockBreakpoint, Functions, FunctionStates, InputState, Message, OutputState, OverallState};
 
 /// A debug handler for interacting between the CLI client and the Debugger in the Coordinator
