+++ flowr/src/cli/flowr_context_functions.md
@@ -4,5 +4,3 @@ interacting with the execution environment.
 
 `flowr` is one particular, Command Line oriented, flow runner, and it provides a set of `context functions` to 
 interact with the file system and standard input/output.
-
-These functions are described in more detail in the following sections:
